+++ Src/Commands/ReadXmlCommand.cs
@@ -1,9 +1,8 @@
 
-// Copyright (c) 2015 Roman Kuzmin
+// Copyright (c) 2015-2016 Roman Kuzmin
 // http://www.apache.org/licenses/LICENSE-2.0
 
 using System.Management.Automation;
-using System.Xml;
 
 namespace Xmlips.Commands
 {
@@ -18,36 +17,12 @@ namespace Xmlips.Commands
 		public string Content { get; set; }
 
 		[Parameter]
-		public SwitchParameter Fragment { get; set; }
-
-		[Parameter]
 		public SwitchParameter Backup { get; set; }
 
 		protected override void BeginProcessing()
 		{
 			Path = GetUnresolvedProviderPathFromPSPath(Path);
-
-			if (Fragment)
-			{
-				var document = new XmlDocument();
-
-				var settings = new XmlReaderSettings();
-				settings.ConformanceLevel = ConformanceLevel.Fragment;
-
-				using (var reader = XmlReader.Create(Path, settings))
-				{
-					XmlNode node;
-					while ((node = document.ReadNode(reader)) != null)
-					{
-						if (node.NodeType == XmlNodeType.Element)
-							WriteObject(node);
-					}
-				}
-			}
-			else
-			{
-				WriteObject(new FileXmlDocument(Path, Content, Backup));
-			}
+			WriteObject(new FileXmlDocument(Path, Content, Backup));
 		}
 	}
 }